+++ main.py
@@ -1,5 +1,0 @@
-from miminions.cli import main
-
-
-if __name__ == "__main__":
-    main()